requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-escpos>=3.1
pydantic>=2.0
httpx>=0.27.0
